+++ src/render/utils/index.js
@@ -1,4 +1,10 @@
+const { classNameCase, formatDescription, internalCase } = require("../../common");
+const { getInlineParseContent } = require("../../schema");
+
 module.exports = {
+  formatDescription,
+  internalCase,
+  classNameCase,
   fmtToJSDocLine: require("./fmtToJSDocLine"),
   _: require("lodash"),
 };